Migrate validation methods and string-convertion methods away from the parser class.
